refactor: split cmd/api/main.go into logical files
Split the 1,732-line god file into focused modules:
- main.go (176 lines) — main(), APIServer struct, server lifecycle
- routes.go (268 lines) — setupMiddleware(), setupRoutes()
- handlers.go (412 lines) — health, status, config, WebSocket handlers
- handlers_trading.go (523 lines) — orders, positions, trading control
- handlers_agents.go (70 lines) — agent CRUD handlers
- broadcast.go (195 lines) — WebSocket broadcast methods
- helpers.go (153 lines) — utilities, middleware helpers

No logic changes — purely file reorganization within package main.
